microbe_yolo_project/requirements.txt
ultralytics>=8.0.200
opencv-python>=4.8
matplotlib>=3.7
numpy>=1.24
pandas>=2.0
PyYAML>=6.0
requests>=2.31
scikit-learn>=1.3
tqdm>=4.66
